Rework add_term_tally_table_header function and add in headings for sampled/unsampled tables
Now called add_text_to_summary_sheets, this function can be used to add the header to the term tally tables as well as add in search strategy/sample headings for the tables where sampling has occurred.

Looked into merging this with add_header_to_sheet but decided not to as that function is added to both data and summary sheets whereas this just applies to summary sheets. Can potentially still look into merging functions.